Moves removeFields to the views
Behaviour from this bit of code needs to be different depending on the
view that calls it, so I've moved it out of the controller and adapted
the behaviour.

On the sidebar I need to set the input to zero before re-submitting the
form to calculate the emissions correctly.

Please have a look to see if I didn't mess it up!